app/requirements.txt
fastapi
uvicorn[standard]
asyncpg
SQLAlchemy
sqlalchemy_utils
alembic
pydantic
pydantic[email]
pydantic-settings
python-jose
requests
psycopg2-binary
python-dotenv
python-multipart
nanoid
pytz
itsdangerous
cryptography